chore: reorganize channel ugprades tests + add them to compatibility tests
