+++ .github/compatibility-test-matrices/main/ica-channel-upgrade-chain-a.json
@@ -1,0 +1,19 @@
+{
+  "chain-a": [
+    "main"
+  ],
+  "chain-b": [
+    "main",
+    "release-v8.1.x"
+  ],
+  "entrypoint": [
+    "TestInterchainAccountsChannelUpgradesTestSuite"
+  ],
+  "test": [
+    "TestMsgSendTx_SuccessfulTransfer_AfterUpgradingOrdertoUnordered",
+    "TestChannelUpgrade_ICAChannelClosesAfterTimeout_Succeeds"
+  ],
+  "relayer-type": [
+    "hermes"
+  ]
+}
