+++ .github/compatibility-test-matrices/main/ica-channel-upgrade-chain-b.json
@@ -1,0 +1,19 @@
+{
+  "chain-a": [
+    "main",
+    "release-v8.1.x"
+  ],
+  "chain-b": [
+    "main"
+  ],
+  "entrypoint": [
+    "TestInterchainAccountsChannelUpgradesTestSuite"
+  ],
+  "test": [
+    "TestMsgSendTx_SuccessfulTransfer_AfterUpgradingOrdertoUnordered",
+    "TestChannelUpgrade_ICAChannelClosesAfterTimeout_Succeeds"
+  ],
+  "relayer-type": [
+    "hermes"
+  ]
+}
