+++ .github/compatibility-test-matrices/main/transfer-channel-upgrade-chain-b.json
@@ -1,0 +1,19 @@
+{
+  "chain-a": [
+    "main",
+    "release-v8.1.x"
+  ],
+  "chain-b": [
+    "main"
+  ],
+  "entrypoint": [
+    "TestTransferChannelUpgradesTestSuite"
+  ],
+  "test": [
+    "TestChannelUpgrade_WithFeeMiddleware_Succeeds",
+    "TestChannelUpgrade_WithFeeMiddleware_FailsWithTimeoutOnAck"
+  ],
+  "relayer-type": [
+    "hermes"
+  ]
+}
