+++ .github/compatibility-test-matrices/unreleased/transfer-channel-upgrade-chain-b.json
@@ -1,0 +1,18 @@
+{
+  "chain-a": [
+    "release-v8.1.x"
+  ],
+  "chain-b": [
+    "release-v8.1.x"
+  ],
+  "entrypoint": [
+    "TestTransferChannelUpgradesTestSuite"
+  ],
+  "test": [
+    "TestChannelUpgrade_WithFeeMiddleware_Succeeds",
+    "TestChannelUpgrade_WithFeeMiddleware_FailsWithTimeoutOnAck"
+  ],
+  "relayer-type": [
+    "hermes"
+  ]
+}
